+HelpScrollDistance dependency property
This describes how far the UI should scroll when displaying the help
UI. This way we can set different values depending on whether the UI is
expanded. Then the help animation storyboard can bind to this property
and always move the correct distance, eliminating the need for multiple
storyboards geared for different window heights.