Extracting a Confirm Exit method for showing the dialog
We'll isolate this guy and return a true/false depending on whether the
user wants to really exit. The two different code paths (Esc, Alt+F4)
will need to react differently, so they'll call this themselves and do
whatever they need to. (The Closing handler needs to Cancel the event
if the user decides to not quit.)