+CommitKeyPressBehavior blank behavior
This will be the master behavior that governs all the keypresses that occur in the commit dialog. Based on the current app state, it'll send inputs to where they need to be. This is true if the "Confirm Exit" panel is visible, where "Y" and "N" keypresses have a different meaning.