+Generic form of the AsyncEventHandler that returns something
This kind of goes against the usual form for event handler delegate types. I could've derived a new EventArgs class to let the event responder set the "args" instance to a result. This felt kinda weird, since you could just as easily not set a result. With a return type though, you HAVE to do something, which is clearly non-optional.