RebasePanel is now a UserControl, instead of ListBox
This is so we can use another element, a Rectangle, to serve as the thing that highlights the current item. We'll do this instead of using the regular focus style because controlling an external rectangle gives us a lot more flexibility in how we display it.

This commit probably doesn't build either.