Now has a nice way of changing an item's rebase action
This will figure out which text box to fade in, and which to fade out. The one fading out will be animated as such, and the other is faded in with the new action's text in it. Then they're marked as swapped, so next animation will bring the old one around. This should prevent flickering, since controls won't jump around.