CommitDocument is now a parameter into CommitViewModel
This moves the view model instantiation to happen as late as possible. The ViewModelLocator, which gets involved as the DataContext binding occurs, now creates the instance. It resolves the commit document (created by the startup procedure), as well as the view service (which happens to be the WindowBase, which exists only as a window is created)

This also removes the need for the static property for commit document, since now it's delivered via DI to the view model creation. From there, it's a property of the CommitViewModel.