Now pre-registering the CommitViewModel factory method
This stashes the creation method, not the actual instance. This is important, since at run time it will actually call the factory method, which will do its own resolves. Because this happens so late, all the resolves will succeed, because all the resources will exist. This isn't true on app launch, which we why we don't stash the INSTANCE.

Pulling the commit file reading out of the app controller. This makes AppController more single-purpose, with the path-switching logic occur in App.xaml.cs. Not ideal, but makes it more uniform.