Now removes, animates, and sets a new adorner when changing indices
After the adorner is moved, a rectangle control is added in its place
with an animation to move it (via margin) to the location of the next
item. When it arrives, the rectangle is removed and an adorner is added
to the new item. This makes it look like the selection rectangle moves
between them fluidly, when really it's going through a few controls.