+Overlay grid for exit animation
This will gradually become visible and animate in from the left, scaling to the right, to occupy the entire space in the text box. It uses a secondary TextBox with a blur effect to obscure the content underneath. Only putting a blur on this border doesn't blur it ENOUGH. So instead, we'll gradually REPLACE the underlying TextBox with the blurred TextBox. This is where the animated Clip geometries come in: As the blured TextBox clips in, the regular one under the green overlay will clip out.