Secondary notes tray now fades in along with expanding downward
This gives the impression that it comes from nowhere, even though it's there the entire time. This seems cool, but having a sliver of it always visible seems like it could be a UI affordance, and maybe a thing to click to enable some mouse control?

We'll try this for now and see how it goes.