MainEntryBox now focuses its text box when it receives focus
This ensures that focusing the control itself will direct the focus to the text box--the thing you actually care about having focus.

Note that this works with the main startup logic, focusing the MainEntryBox on launch, since he'll turn around and focus the text box, allowing you to type after launch.