-ExitReason converters
We're only using this now in the Transition box, which we do in the code-behind. Because we instantiate this guy outside the visual tree, the converters don't have a chance to fire, so we hard-code the logic. We could probably move it out into a "decider" kind of class that we can unit test, but do that later.