Making ExpandUI() a real async method
This oughta fix the broken test, where it couldn't be awaited, causing early returns and failing the test.

Also commenting out some tests. What we really need is an async command that won't yield to its caller on the first await. For now it's a wrench in the gears around smart-paste, so we're commenting out the tests for now, since the real functionality is also disabled.