Now drawing the primary commit notes on the "front material"
This ensures as we exit, whatever got typed in is appears on the rotating UI, instead of going blank (where we swap the real UI for the fake material). Note that the radial counter still doesn't appear correctly and will revert to 72, since that's the default on the material we started with.