Making expand/collapse animations a little smoother
We do this by keeping the window height the expanded size the whole time, and instead just growing the border control, rather than having to animate the height of the window.

Also re-arranged the control that got animated. This lets the height go from 0 -> max, with no partial control hanging out. This ALSO lets us ditch the opacity animation, which probably wasn't helping the performance.