Displays an error dialog if it can't load the incoming file
This is either because it doesn't exist or because it's 0 bytes. The dialog also grabs the "main window" ahead of time, because if there's no main window, it somehow tries to set its owner to itself. So now we'll grab it ahead of time (which is null), and then set a null owner.