+SuppressAlt behavior, logic controller, and test classes
This is a thing that will attach to a window and suppress the Alt key from being processed. This means that pressing Alt will NOT activate the "accelerator selection" input state for any window it's attached to. The "logic controller" is the thing that really makes the calls, where the behavior itself should defer decisions to the logic controller.