Now calls "animate" on the entry box when Ctrl is pressed / released
Added a new glyph (right arrow) for it, to imply accept. Didn't want to use the check, but instead wanted to use a more verb-oriented icon. The check seems like an end state you've entered, not a way to get there.

The "animate" method doesn't do anything yet. It'll soon hand it down to the right control to do the actual animation.