Button flashes now originate from the click location on the button
The button content is now a Canvas instead of a Grid. This lets us animate the Canvas.Left and Canvas.Top to place (and animate) the flash as it plays. The Ellipse is anchored by its top-left coordinate, not its center, so as it expands we need to animate its position to keep it centered.

Since it's a canvas, the AccessText isn't auto-centered anymore, but we stuck it in a Grid, matching that size to the button's. It's pretty close, but needs a tiny hard-coded adjustment. It looks all right.

This sure ain't pretty, but it works.