Fixing warning about async test; more smartly raising async events
Using null conditional isn't good for event handlers. When the handler is null, we end up awaiting it, which yields a null task, which is a null reference exception.

Here we add a RaiseAsync method and pass the event as a Delegate instance. Somehow it knows it's an AsyncEventHandler--I have no idea how this works, but it seems to. We can pass the event around and have it safely raised elsewhere.