Splitting RaiseAsync into two methods for Shutdown
This hid a bug where the delegate type was AsyncEventHandler<>, which failed the "as" cast, which returned null, which returned a completed task, instead of invoking the event handler.

It seemed impossible to define this parameter generically to encapsulate what kind of delegate it is. Instead, we RaiseAsync into two methods, one that takes the AsyncEventHandler, and the other that takes the templated one.