Fixing the "multi-constructor" problem for CommitFileReader
Instead of registering in IOC with a factory, we now map to the type instead--this lets it call the constructor itself. This previously was a problem because there were two constructors--default and with dependencies. By eliminating the default constructor, the IOC can create it easily, and we're still able to pass dependencies in unit tests.