+Margin for the primary text box for commit notes
As text hangs off the edge this looks much better, since it forces the text to be wholly visible, instead of hidden behind the radial counter.

This happens when editing/amending a commit whose text is above the limit, so you'll get a free pass on the length for that commit.