Now generates materials after the UI has loaded
This ensures the startup animation is nice and smooth, so startup time isn't affected by this process. When done concurrently, it did make the animation choppy, since it was doing so much.

Now, we wait til the window has fully loaded to create the materials. This meant that you could also exit the app before the materials had finished generating. Now the exit process awaits the material generation task, so if you exit quickly, you may have to wait a tiny amount for the materials to finish. This is pretty unnoticeable, and you'd have to exit basically immediately after the app started; this case isn't practical, but DOES prevent a crash, which is why it's in there.