+MessageReceiver generic type
We'll try out the notion of a "receiver," which is a distributed, unrelated object that acts sort of as a "view model" for a messenger message. This allows objects to "fire and forget" a message (with data if necessary) and someone else can receive it (receivers), and then glob together whatever they need to to process it.

We should be able to have view models with fewer dependencies, since they're handing it off to collaborators faster, without having to know who they are.