+IMessageReceiver
This is designed to be a non-generic parent that we can retain instances of, that really are generic receivers. It also lets us pass a messenger to it and force registration. This undoes the current pattern of automatic registration in the constructor. We need a hook to latch onto if we want to replace the messenger with a different one in test.