+ReceiverHarness test class to smooth out testing receivers
This guy will automatically create and register a new messenger to a receiver instance. From there you can pipe a message to it so the receiver's OnReceive method is called (the real system under test). The harness can also "Verify" that messages were sent (similar to a Moq Verify). This is useful when the receiver-under-test goes on to send other messages as part of its implementation.