Fixing a busted test that was asserting the wrong thing
Now asserts that the ExitApplicationMessage was sent. This surfaced a problem that we were calling the wrong form of Send() on IMessenger, since the sub-class was hiding it with the "new" keyword. To solve this, I removed the type constraint and no longer keep track of MessageBase, but now just regular object. Verify() queries are the same.