First pass at a string-to-string-array converter
We'll strap this to the "body" binding. GitModel gives us an array of strings as the body, instead of one big string, and we're binding to a text box, which is one big string. So this converter will help us go between both without logic in the view model.

This first part translates from a string[] -> string (pushing into the text box from the view model). The next part will go the other direction--pushing into the model as the UI changes.