+CommitScenarioBuilder
This is the thing that sets all everything in place to actually create the CommitScenario. This figures out the default object composer, but can OVERRIDE mappings (it does this to replace IViewService with a mock, so we don't actually try to show dialogs).

This class ultimately produces the CommitScenario that's been configured and is ready to go.