Foregrounds are now read directly from settings with change ability
I couldn't figure out how to implement these in the Style, since they seem to get evaluated first. Got cryptic runtime exceptions. Decided to punt it for now and require that if you're using a property from settings, it has to be done in the actual XAML, and not a style. This works great, but could probably be made cleaner with less duplication. For now you have to "use it right."