SettingsExtension can now use a converter when providing a value
This lets it source values from settings, but lets consumers add special behavior if they need. We'll use this to simulate the accent color--primarily use the foreground color, but attach the ColorOpacityConverter to slightly fade it. It's wide open for general use on anything, but this'll be a big use.