Unifying save and discard paths under the Deactive path
This means we only save in one place and are in a good spot for verification on discard with pending changes. This uses a Func to provide the document we're going to save. It defaults to empty (for the discard case), and on Save it'll rework this to be the real document to save.